requirements.txt
aiohttp
discord-webhooks
discord.py
python-dotenv
